packages/013_fluence-ocr/requirements.txt
flask==3.0.0
pillow==10.1.0
pytesseract==0.3.10
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3
